/input 모듈화
